+++ examples/InteractiveBrokersExamples/src/main/java/com/sumzerotrading/ib/example/trading/TradingExample.java
@@ -27,6 +27,9 @@ import com.sumzerotrading.data.FuturesTicker;
 import com.sumzerotrading.data.StockTicker;
 import com.sumzerotrading.interactive.brokers.client.InteractiveBrokersClient;
 import com.sumzerotrading.interactive.brokers.client.InteractiveBrokersClientInterface;
+
+import java.math.BigDecimal;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -56,9 +59,9 @@ public class TradingExample {
         int contracts = 5;
 
         // Create the order and send to Interactive Brokers
-        TradeOrder order = new TradeOrder(orderId, futuresTicker, contracts, TradeDirection.BUY);
+        TradeOrder order = new TradeOrder(orderId, futuresTicker, BigDecimal.valueOf(contracts), TradeDirection.BUY);
         order.setType(TradeOrder.Type.LIMIT);
-        order.setLimitPrice(32.50);
+        order.setLimitPrice(BigDecimal.valueOf(32.50));
         ibClient.placeOrder(order);
     }
 
@@ -68,7 +71,7 @@ public class TradingExample {
         String orderId = ibClient.getNextOrderId();
         int shares = 500;
 
-        TradeOrder order = new TradeOrder(orderId, amazonTicker, shares, TradeDirection.SELL);
+        TradeOrder order = new TradeOrder(orderId, amazonTicker, BigDecimal.valueOf(shares), TradeDirection.SELL);
 
         // ibClient.placeOrder(order);
         ibClient.getOpenPositions();
@@ -86,7 +89,7 @@ public class TradingExample {
         logger.info("Next Order ID: " + orderId);
         int amount = 50000;
 
-        TradeOrder order = new TradeOrder(orderId, eurTicker, amount, TradeDirection.BUY);
+        TradeOrder order = new TradeOrder(orderId, eurTicker, BigDecimal.valueOf(amount), TradeDirection.BUY);
 
         logger.info("Waiting for 15 seconds before placing order...");
         Thread.sleep(20000);
